% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cache.R
\name{fastmart_annotate_biotype}
\alias{fastmart_annotate_biotype}
\title{Annotate ENSEMBL GENES with BIOTYPE}
\usage{
fastmart_annotate_biotype(
  ensemble_gene_id,
  GRCh = c("38", "37"),
  cache_dir = fastmart_default_cache()
)
}
\arguments{
\item{ensemble_gene_id}{ensembl gene id (can be versioned / not versioned)}

\item{GRCh}{assembly to use. Can only be set to}

\item{cache_dir}{directory to store cache in}
}
\value{
matched biotype for each sample (character)
}
\description{
Annotate ENSEMBL GENES with BIOTYPE
}
\examples{
\dontrun{
fastmart_annotate_biotype('ENSG00000282455.1')
#> IG_D_gene
}
}
