% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cache.R
\name{fastmart_cache_tables}
\alias{fastmart_cache_tables}
\title{Cache Tables}
\usage{
fastmart_cache_tables(
  ensembl_version = 109,
  overwrite = FALSE,
  GRCh = c("38", "37"),
  cache_dir = fastmart_default_cache()
)
}
\arguments{
\item{ensembl_version}{ensembl version. Older versions may not be supported. Will be ignored if GRCh != 38}

\item{overwrite}{overwrite existing cache (flag)}

\item{GRCh}{assembly to use. Can only be set to}

\item{cache_dir}{directory to store cache in}
}
\value{
invisible(TRUE). This function is run for its side effects
}
\description{
Cache Tables
}
\examples{
\dontrun{
fastmart_cache_tables()
}
}
