% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cache.R
\name{fastmart_convert_hgnc_to_ensembl}
\alias{fastmart_convert_hgnc_to_ensembl}
\title{Convert HGNC -> ENSEMBL IDs}
\usage{
fastmart_convert_hgnc_to_ensembl(
  hgnc_symbols,
  chrom,
  start,
  end,
  GRCh = c("38", "37"),
  exact_match = TRUE,
  cache_dir = fastmart_default_cache()
)
}
\arguments{
\item{hgnc_symbols}{HGNC symbols}

\item{chrom}{Chromosome of Gene. Required to resolve ambiguous mappings (character)}

\item{start}{Start position of Gene. Required to resolve ambiguous mappings  (numeric)}

\item{end}{End position of Gene. Required to resolve ambiguous mappings  (numeric)}

\item{GRCh}{assembly to use. Can only be set to}

\item{exact_match}{Is an exact match between chrom-start-end required (TRUE), or should the omst similar interval be returned (FALSE)  (flag)}

\item{cache_dir}{directory to store cache in}
}
\value{
matched ENSEMBL IDs (character)
}
\description{
Because ENSEMBL IDs are more granular than HGNC / HUGO symbols we require the start & stop coordinates
}
\examples{
\dontrun{
fastmart_convert_hgnc_to_ensembl('TBCE', chrom='1', start='235530675', end='235612283', GRCh = "37")
# > ENSG00000116957.8
}

}
