% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cache.R
\name{fastmart_default_cache}
\alias{fastmart_default_cache}
\title{Default Cache}
\usage{
fastmart_default_cache()
}
\value{
default path to cache
}
\description{
Default Cache
}
\examples{
fastmart_default_cache()
}
